refactor(download): extract shared net module for binary acquisition
Collapse the duplicated parallel-download and HTTP-fetch logic that
agent_binary and capsule_binary each carried into one src/net module.
http_client, get_text/fetch_text, and download_parallel now live in a
single place, so the fast-down tuning constants and the GET-to-String
shape have one home instead of two drifting copies.

- Per-chunk download errors now log on the debug_log! (cdebug!) tier in
  the shared helper, matching the two-tier telemetry rule; they were on
  agent_binary's always-on tier before.
- Factor the cache-hit -> download -> ready tail in ensure_available into
  ensure_binary_for_release, shared by the metadata-cache and
  post-resolve paths.
- Route the GitHub API retry closure through net::get_text instead of
  re-inlining the GET body.
- Drop a dead `let expected_sha = expected_sha;` self-rebind and a
  redundant typed rebind of the agent-binaries vec.

No behavior change beyond the telemetry-tier move. 2238 lib tests pass.
